replaced no types examples with original for standalone
