+++ examples/kokkos-benchmarks/standalone/gups.py
@@ -5,23 +5,20 @@ from typing import Tuple
 import pykokkos as pk
 
 @pk.workunit
-def init_data(i: int, data: pk.View1D[pk.int64]):
+def init_data(i, data):
     data[i] = 10101010101
 
 @pk.workunit
-def init_indices(i: int, indices: pk.View1D[pk.int64]):
+def init_indices(i, indices):
     indices[i] = 0
 
 @pk.workunit
-def run_gups_atomic(i: int, data: pk.View1D[pk.int64], \
-        indices: pk.View1D[pk.int64], datum: pk.int64):
+def run_gups_atomic(i, data, indices, datum):
     pk.atomic_fetch_xor(data, [indices[i]], datum)
 
 @pk.workunit
-def run_gups(i: int, data: pk.View1D[pk.int64], \
-        indices: pk.View1D[pk.int64], datum: pk.int64):
+def run_gups(i, data, indices, datum):
     data[indices[i]] ^= datum
-
 
 if __name__ == "__main__":
     random.seed(1010101)
